requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.8
dataclasses==0.6
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
Django==5.1.1
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
dnspython==2.6.1
elastic-transport==8.15.0
elasticsearch==8.15.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==0.2.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
huggingface-hub==0.25.2
idna==2.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
langcodes==3.4.0
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
postgrest-py==0.4.0
preshed==3.0.9
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
py==1.11.0
py4j==0.10.9.7
pycurl==7.45.3
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyotp==2.9.0
pyparsing==3.1.2
pyspark==3.5.2
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.1.0
realtime-py==0.1.3
regex==2024.9.11
requests==2.25.1
rfc3986==1.5.0
rich==13.8.1
safetensors==0.4.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
supabase-py==0.0.2
sympy==1.13.3
thinc==8.2.5
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
twikit==2.1.3
twilio==9.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==9.1
Werkzeug==3.0.4
wfuzz==3.1.0
wrapt==1.16.0
yarl==1.11.1
gunicorn
flask_cors
